+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `setRecent` function
 - `getAutoSave` function
 - `setAutoSave` function
+- Markdown preview
+### Changed
+- `service-worker.js` modified
+- `README.md` updated
 ## [0.4] - 2026-04-15
 ### Added
 - Download button